Updated works controller action index to utilize two new written methods for works. Updated work model with two methods self.ordered_works and self.order_by_votes. Updated edit in works controller. Updated update action in works controller with an if-else to choose redirection path. Rearranged links in the category view. Added vote counts to the works index view. Added foundation code to show view for works. Added new test for works-controller create.

+++ app/controllers/works_controller.rb
@@ -1,8 +1,14 @@
 class WorksController < ApplicationController
   def index
-    @albums = Work.where(category: "album").limit(10)
-    @books = Work.where(category: "book").limit(10)
-    @movies = Work.where(category: "movie").limit(10)
+    # @albums = Work.where(category: "album").limit(10)
+    # @books = Work.where(category: "book").limit(10)
+    # @movies = Work.where(category: "movie").limit(10)
+    @albums = Work.ordered_works("album")[0...10]
+    @books = Work.ordered_works("book")[0...10]
+    @movies = Work.ordered_works("movie")[0...10]
+    leader_id = Work.order_by_votes(Work.all)[0]
+    @leader = Work.find_by(id: leader_id)
+
   end
 
   def new
@@ -52,6 +58,9 @@ class WorksController < ApplicationController
 
   def edit
     @work = Work.find_by(id: params[:id])
+    if @work.nil?
+      head :not_found
+    end
     session.delete(:return_to)
     session[:return_to] ||= request.referer
     @back_url = session[:return_to]
@@ -61,7 +70,16 @@ class WorksController < ApplicationController
     work = Work.find_by(id: params[:id])
     work.update_attributes(work_params)
     work.save
-    redirect_to session[:return_to]
+    if session[:category] == "album"
+      path = albums_path
+    elsif session[:category] == "book"
+      path = books_path
+    elsif session[:category] == "movie"
+      path = movies_path
+    end
+    # redirect_to session[:return_to]
+    redirect_to path
+
   end
 
   def destroy
